requirements.txt
hydra-core==1.3.2
numpy==1.24.1
omegaconf==2.3.0
pandarallel==1.6.5
pandas==1.5.2
python-dotenv==1.0.0
python_telegram_bot==20.1
PyYAML==6.0
scikit_learn==1.3.0
telegram==0.0.1
torch==2.0.1+cu118
tqdm==4.64.1
transformers==4.26.1
wandb==0.14.2
